+++ docs/source/conf.py
@@ -46,9 +46,9 @@ copyright = '2021, S. Domanskyi, M. Zheng, C. Piermarocchi, G. Mias'
 author = 'S. Domanskyi, M. Zheng, C. Piermarocchi, G. Mias'
 
 # The short X.Y version
-version = '1.2.9'
+version = '1.3.0'
 # The full version, including alpha/beta/rc tags
-release = '1.2.9'
+release = '1.3.0'
 
 
 # -- General configuration ---------------------------------------------------